Implement initial repository setup
Created initial directory structure for this application, as well as some initial files that will be needed as each portion of the application is built.  Client folder contains a src folder for all React component files that Webpack will compile into a bundle.js file in the client's dist directory.