Implement routes and controllers for all requests
Created 2 routes in the routes.js file: 1 for seeding the database with information from Spotify's API (/seedDatabase), and 1 for all user interactions in the application: browsing through cards, posting a new card to the server which is inserted into the database, and deleting a card from the database.  All controllers currently contain console.logs so each function and endpoint can be tested through Postman.