Expand getNextCards function to handle deletes
Added a conditional to getNextCards to ensure that the function will always return 10 cards (in sequence, with no repeats), even if a card has been deleted.  If the return from the database is less than ten, an additional pull from the database will return the correct number of cards.  Otherwise, if the first request returns 10 cards, it will return the cards to the client the first time.