Finish postNewCard function
When a user creates a newCard, a POST request is sent, and getNextCards uses .create() to create a new document in the database. A counter keeps track of the last id in the database - when a new document is entered, the counter increments, ensuring new cards will be returned by subsequent GET requests.